requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.6.0b9
dataclasses-json==0.6.4
distro==1.9.0
docstring-inheritance==2.1.2
duckduckgo_search==4.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.18
langchain-core==0.1.19
langchain-experimental==0.0.50
langsmith==0.0.87
lxml==5.1.0
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.11.1
packaging==23.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
